feat : add confirmation dialog box

+++ src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -55,16 +56,30 @@ export class CoursesComponent implements OnInit {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe({
-      next: () => {
-        this.refresh();
-        this.snackBar.open("Curso removido com sucesso!", 'X', { duration: 5000 ,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',},
-        )
-      },
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
 
-      error: () => this.onError('Erro ao tentar remover curso!'),
-  });
+    dialogRef.afterClosed().subscribe({
+      next: (result: boolean) => {
+
+        if (result) {
+          
+          this.coursesService.remove(course._id).subscribe({
+            next: () => {
+              this.refresh();
+              this.snackBar.open("Curso removido com sucesso!", 'X', { duration: 5000 ,
+                verticalPosition: 'top',
+                horizontalPosition: 'center',},
+              )
+            },
+
+            error: () => this.onError('Erro ao tentar remover curso!'),
+        });
+        }
+      }
+    });
+
+
   }
 }
